Moved remap stuff outside of field folder, plus some mods
The abstract remapper now does not have a default constructor. One needs
to provide source and target grids.

I also removed some "utilities" from remap_utils, since they seemed pointless.
The infrastructure of the driver+field_repo should already be able to
detect remapping types.
